+++ examples/readme_echo_server.rs
@@ -10,11 +10,12 @@ fn main() {
     network.listen(Transport::Ws, "0.0.0.0:3044").unwrap(); //WebSockets
 
     loop {
-        match events.receive() { // Read the next event or wait until have it.
+        match events.receive() {
+            // Read the next event or wait until have it.
             NetEvent::Message(endpoint, data) => {
                 println!("Received: {}", String::from_utf8_lossy(&data));
                 network.send(endpoint, &data);
-            },
+            }
             NetEvent::Connected(_endpoint) => println!("Client connected"), // Tcp or Ws
             NetEvent::Disconnected(_endpoint) => println!("Client disconnected"), //Tcp or Ws
         }